一、修改下单及订单列表 1.下单前判断某包房某时间是否已预订. 2.下单时,添加订单明细 2.订单明细添加预订关联,以便列表查询时显示预订信息 二、添加三级分销模块内容 1.加入注册成为分销职员 2.创建订单时判断是否有分销职员,有分销职员则加入分销职员ID 3.分销职员使用相关功能:我的下级分销人员、通过我的分享链接购买的订单 4.基类控制器,查询是否有分销职员ID参数传递.

+++ app/modules/api/classes/controller/order.php
@@ -84,10 +84,14 @@ class Controller_Order extends Controller_BaseController {
             $item->created_at = date('Y-m-d H:i', $item->created_at);
             $item->order_label = \Model_Order::$_maps['labels'][$item->order_status];
             $item->order_status_title = \Model_Order::$_maps['status'][$item->order_status];
-            $item->details;
             foreach ($item->details as $detail){
-                $detail->goods->title = \Str::truncate($detail->goods->title, 25, '...');
-
+                if($item->order_type == 'RESERVE'){
+                    $detail->reserve->room;
+                    $detail->reserve->reserve_date = date('Y年m月d日', $detail->reserve->begin_at);
+                    $detail->reserve->reserve_time = date('H:i', $detail->reserve->begin_at) . ' 至 ' . date('H:i', $detail->reserve->end_at);
+                }else{
+                    $detail->goods->title = \Str::truncate($detail->goods->title, 25, '...');
+                }
             }
         }
